refactor: 💡 isolate handler in its own file
